test(webauthn): extract SuppressesWebauthnDeprecations trait
